+++ docs/v3/introduction/changelog.md
@@ -3,13 +3,17 @@ title: Changelog
 ---
 
 ## 3.2 - The one with NextJS support
+
 ::: warning Work in progress
 This update is currently being worked on, but will be released soon!
 :::
+
 - Added [NextJS support]() for server-side rendering
 - Added default Group for Collections [More]()
+- Removed Herobrine.
 
 ## 3.1 - The one with Events
+
 ::: tip NPM version
 This is the current release build available on NPM
 :::
